Improved compass and added points selection by azimuth.

+++ app/src/test/java/com/louisnard/augmentedreality/PointTest.java
@@ -24,48 +24,48 @@ public class PointTest {
      */
     @Test
     public void test_degreesToMeters_metersToDegrees() {
-        double angle = 0;
-        double distance = 0;
+        float angle = 0;
+        int distance = 0;
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
 
         angle = 90;
-        distance = 2 * Math.PI * Point.EARTH_RADIUS / 4;
+        distance = (int) (2 * Math.PI * Point.EARTH_RADIUS / 4);
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
 
         angle = 180;
-        distance = 2 * Math.PI * Point.EARTH_RADIUS / 2;
+        distance = (int) (2 * Math.PI * Point.EARTH_RADIUS / 2);
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
 
         angle = 360;
-        distance = 2 * Math.PI * Point.EARTH_RADIUS;
+        distance = (int) (2 * Math.PI * Point.EARTH_RADIUS);
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
 
         angle = -90;
-        distance = 2 * Math.PI * Point.EARTH_RADIUS / 4;
+        distance = (int) (2 * Math.PI * Point.EARTH_RADIUS / 4);
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
 
         angle = -180;
-        distance = 2 * Math.PI * Point.EARTH_RADIUS / 2;
+        distance = (int) (2 * Math.PI * Point.EARTH_RADIUS / 2);
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
 
         angle = -360;
-        distance = 2 * Math.PI * Point.EARTH_RADIUS;
+        distance = (int) (2 * Math.PI * Point.EARTH_RADIUS);
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
 
         angle = 1;
-        distance = 2 * Math.PI * Point.EARTH_RADIUS / 360;
+        distance = (int) (2 * Math.PI * Point.EARTH_RADIUS / 360);
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
 
         angle = -1;
-        distance = 2 * Math.PI * Point.EARTH_RADIUS / 360;
+        distance = (int) (2 * Math.PI * Point.EARTH_RADIUS / 360);
         assertEquals(distance, Point.degreesToMeters(angle), ERROR_TOLERANCE * distance);
         assertEquals(angle, Point.metersToDegrees(distance), ERROR_TOLERANCE * distance);
     }
